+++ testmod-common/src/main/java/io/github/jamalam360/testmod/Flags.java
@@ -1,5 +1,7 @@
 package io.github.jamalam360.testmod;
 
 public class Flags {
+	public static final boolean DEBUG_INTERACTION_EVENTS = false;
 	public static final boolean DEBUG_MOUSE_SCROLL_EVENTS = false;
+	public static final boolean DEBUG_CONTAINER_RENDER_EVENTS = true;
 }